Clean and add comments
